docs: organize text components into group
